refactor: move actions to internals
