+++ ulauncher/api/shared/action/ActionList.py
@@ -1,7 +1,1 @@
-from __future__ import annotations
-
-from typing import Any
-
-
-def ActionList(actions: list[Any]) -> dict[str, Any]:
-    return {"type": "action:legacy_run_many", "data": actions}
+from ulauncher.internals.actions import ActionList  # noqa: F401
