+++ ulauncher/api/shared/action/CopyToClipboardAction.py
@@ -1,5 +1,1 @@
-from __future__ import annotations
-
-
-def CopyToClipboardAction(text: str) -> dict[str, str]:
-    return {"type": "action:clipboard_store", "data": text}
+from ulauncher.internals.actions import Copy as CopyToClipboardAction  # noqa: F401
