+++ ulauncher/api/shared/action/OpenAction.py
@@ -1,5 +1,1 @@
-from __future__ import annotations
-
-
-def OpenAction(item: str) -> dict[str, str]:
-    return {"type": "action:open", "data": item}
+from ulauncher.internals.actions import Open as OpenAction  # noqa: F401
