+++ ulauncher/api/shared/action/RunScriptAction.py
@@ -1,5 +1,1 @@
-from __future__ import annotations
-
-
-def RunScriptAction(script: str, args: str = "") -> dict[str, str | list[str]]:
-    return {"type": "action:legacy_run_script", "data": [script, args]}
+from ulauncher.internals.actions import RunScript as RunScriptAction  # noqa: F401
