+++ ulauncher/modes/file_browser/CopyPathToClipboardResult.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from ulauncher.api.shared.action.CopyToClipboardAction import CopyToClipboardAction
+from ulauncher.internals import actions
 from ulauncher.internals.query import Query
 from ulauncher.internals.result import Result
 
@@ -11,4 +11,4 @@ class CopyPathToClipboardResult(Result):
     icon = "edit-copy"
 
     def on_activation(self, _query: Query, _alt: bool = False) -> dict[str, str]:
-        return CopyToClipboardAction(self.path)
+        return actions.Copy(self.path)
