+++ ulauncher/modes/file_browser/FileBrowserResult.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from os.path import basename, dirname, isdir, join
 
-from ulauncher.api.shared.action.OpenAction import OpenAction
+from ulauncher.internals import actions
 from ulauncher.internals.query import Query
 from ulauncher.internals.result import Result
 from ulauncher.modes.file_browser.CopyPathToClipboardResult import CopyPathToClipboardResult
@@ -31,7 +31,7 @@ class FileBrowserResult(Result):
             if isdir(self.path):
                 return join(fold_user_path(self.path), "")
 
-            return OpenAction(self.path)
+            return actions.Open(self.path)
 
         if isdir(self.path):
             open_folder = OpenFolderResult(name=f'Open Folder "{basename(self.path)}"', path=self.path)
