+++ ulauncher/modes/file_browser/OpenFolderResult.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from ulauncher.api.shared.action.OpenAction import OpenAction
+from ulauncher.internals import actions
 from ulauncher.internals.query import Query
 from ulauncher.internals.result import Result
 
@@ -11,4 +11,4 @@ class OpenFolderResult(Result):
     path = ""
 
     def on_activation(self, _query: Query, _alt: bool = False) -> dict[str, str]:
-        return OpenAction(self.path)
+        return actions.Open(self.path)
